requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosmtplib==3.0.1
aiosqlite==0.19.0
alembic==1.12.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
asttokens==2.4.1
attrs==23.1.0
black==23.12.1
certifi==2023.11.17
click==8.1.7
decorator==5.1.1
executing==2.0.1
fastapi==0.108.0
flake8==6.1.0
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
ipython==8.18.1
jedi==0.19.1
magic-filter==1.0.12
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
Pygments==2.17.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.32.0.post1
traitlets==5.14.0
typing_extensions==4.8.0
tzlocal==5.2
uvicorn==0.25.0
wcwidth==0.2.12
yarl==1.9.3
